WIP: move IPA srs into evaluation_proof (soon to be renamed into ipa)
Also remove duplicated method called max_degree, which was an alias of
max_poly_size
Same for add_lagrange_basis and create.
