+++ kimchi/src/circuits/expr.rs
@@ -3539,8 +3539,8 @@ pub mod test {
     use ark_ff::UniformRand;
     use mina_curves::pasta::{Fp, Pallas, Vesta};
     use poly_commitment::{
-        evaluation_proof::OpeningProof,
-        srs::{endos, SRS},
+        evaluation_proof::{endos, OpeningProof, SRS},
+        SRS as _,
     };
     use rand::{prelude::StdRng, SeedableRng};
     use std::{array, sync::Arc};
