+++ kimchi/src/tests/rot.rs
@@ -28,8 +28,8 @@ use mina_poseidon::{
 };
 use o1_utils::Two;
 use poly_commitment::{
-    evaluation_proof::OpeningProof,
-    srs::{endos, SRS},
+    evaluation_proof::{endos, OpeningProof, SRS},
+    SRS as _,
 };
 use rand::Rng;
