+++ msm/src/logup.rs
@@ -139,9 +139,8 @@
 //! - evaluate all polynomials at the evaluation points ζ and ζω (because we
 //! access the "next" row for the accumulator in the quotient polynomial).
 use ark_ff::{Field, PrimeField, Zero};
-use std::{collections::BTreeMap, hash::Hash};
-
 use kimchi::circuits::expr::{BerkeleyChallengeTerm, ConstantExpr, ConstantTerm, Expr, ExprInner};
+use std::{collections::BTreeMap, hash::Hash};
 
 use crate::{
     columns::Column,
