+++ poly-commitment/tests/batch_15_wires.rs
@@ -12,8 +12,7 @@ use mina_poseidon::{
 use o1_utils::ExtendedDensePolynomial as _;
 use poly_commitment::{
     commitment::{combined_inner_product, BatchEvaluationProof, CommitmentCurve, Evaluation},
-    evaluation_proof::DensePolynomialOrEvaluations,
-    srs::SRS,
+    evaluation_proof::{DensePolynomialOrEvaluations, SRS},
     SRS as _,
 };
 use rand::Rng;
